+++ docs/source/user_guide/changelog.rst
@@ -4,7 +4,7 @@ Changelog
 
 Authors since :ref:`5.0.0`: Jordi Torrents (jordi.torrentsm@gmail.com)
 
-6.0.4
+6.0.5
 =====
 
 - Add a new GUI for the :ref:`video generator`.